requirements.txt
streamlit>=1.32.0
PyPDF2==3.0.1
sounddevice>=0.4.6
numpy>=1.24.0
langchain>=0.0.352
langchain-community>=0.0.10
langchain-openai>=0.0.3
beautifulsoup4>=4.12.2
requests>=2.31.0
pandas>=2.1.0
python-dotenv>=1.0.0
pinecone>=0.8.0
openai>=1.3.0
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
markdown>=3.5.1
pytest>=7.4.3
python-multipart>=0.0.6
aiohttp>=3.9.1
tiktoken>=0.5.1
rich>=13.7.0
protobuf>=5.29,<6.0